requirement.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
sqlalchemy
psycopg2
#-e